LIME-1748 - Moved Stub and VC Validation steps to common
Steps for Stub login and VC validation have been moved to a specific StepDef and Page file
Removed method which are no longer in use
Tidied up methods in the wrong page file
Resolved Sonar errors
