--- acceptance-tests/src/test/java/gov/di_ipv_drivingpermit/model/AddressType.java
@@ -1,6 +1,0 @@
-package gov.di_ipv_drivingpermit.model;
-
-public enum AddressType {
-    CURRENT,
-    PREVIOUS
-}
