+++ acceptance-tests/src/test/java/gov/di_ipv_drivingpermit/pages/GlobalPage.java
@@ -5,14 +5,7 @@ import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 
-import java.util.HashMap;
-import java.util.Map;
-
 public class GlobalPage extends PageObjectSupport {
-
-    static final By CONTINUE_BUTTON = By.xpath("//button[@class='govuk-button button']");
-
-    private static Map<String, String> jsonFileResponses = new HashMap<>();
 
     WebDriver driver;
