+++ acceptance-tests/src/test/java/gov/di_ipv_drivingpermit/pages/Headers.java
@@ -2,13 +2,5 @@ package gov.di_ipv_drivingpermit.pages;
 
 public class Headers {
 
-    public static final String CHECKING_YOUR_DETAILS = "We need to check your details";
     public static final String IPV_CORE_STUB = "IPV Core Stub - GOV.UK";
-    public static final String CREDENTIAL_ISSUERS = "Visit Credential Issuers";
-    public static final String PROBLEM_WITH_THE_SERVICE =
-            "Sorry, there is a problem with the service";
-    public static final String DL_CRI_ENVIRONMENT = "User for driving licence CRI";
-    public static final String VERIFIABLE_CREDENTIALS = "Verifiable Credentials";
-    public static final String THERE_IS_A_PROBLEM = "There is a problem.";
-    public static final String ORCHESTRATOR_STUB = "Orchestrator Stub";
 }
