--- acceptance-tests/src/test/java/gov/di_ipv_drivingpermit/pages/ProveYourIdentityGovUkPage.java
@@ -1,11 +1,0 @@
-package gov.di_ipv_drivingpermit.pages;
-
-import org.openqa.selenium.By;
-
-public class ProveYourIdentityGovUkPage extends GlobalPage {
-    private static final By USER_INFO = By.cssSelector(".govuk-heading-l");
-
-    public void waitForPageToLoad() {
-        waitForElementVisible(USER_INFO, 30);
-    }
-}
