+++ acceptance-tests/src/test/resources/features/UKDrivingLicence/DVLA/DVLADrivingLicenceAuthSource.feature
@@ -44,7 +44,7 @@ Feature: DVLA Auth Source Driving Licence Test
     And I add a cookie to change the language to English
     Then I check the page title is Was your UK photocard driving licence issued by DVLA or DVA? – GOV.UK One Login
     And I should see DVLA as an option
-    And I click on DVLA radio button and Continue
+    And I click on DVLA radio button and continue
     And I check the page title is Enter your details exactly as they appear on your UK driving licence – GOV.UK One Login
     And I see a form requesting DVLA LicenceNumber
     Given User enters DVLA data as a <DVLADrivingLicenceSubject>
